requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
black==24.10.0
click==8.1.7
colorama==0.4.6
Django==4.2.17
django-rest-framework==0.1.0
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
redis==5.2.1
sqlparse==0.5.2
tomli==2.2.1
typing-extensions==4.12.2
tzdata==2024.2